collector module
